Move God.js comments into doc/DEBUG.md - move console overidde into Utility.js
