+++ lib/God.js
@@ -11,15 +11,7 @@ var cst           = require('../constants.js');
 var pidusage      = require('pidusage');
 var vizion        = require('vizion');
 var debug         = require('debug')('pm2:god');
-// var observe       = require('observe-js');
-
-// Memory leak inspector
-// require('webkit-devtools-agent').start({
-//   port: 9999,
-//   bind_to: '0.0.0.0',
-//   ipc_port: 3333,
-//   verbose: true
-// });
+var Utility       = require('./Utility.js');
 
 /**
  * Override cluster module configuration
@@ -41,68 +33,7 @@ var God = module.exports = {
   })
 };
 
-///////////////////////////////////////////////////////
-// Temporary disabling this because tests don't pass //
-///////////////////////////////////////////////////////
-
-// var obs = new observe.ObjectObserver(God.clusters_db);
-// obs.open(function() {
-//   God.dumpProcessList && God.dumpProcessList(function() {
-//     console.log('Process List Dumped');
-//   });
-// });
-// if (!Object.observe) {
-//   setInterval(Platform.performMicrotaskCheckpoint, 1000);
-// }
-
-/**
- * Hack Global Console Of PM2
- */
-(function hackConsole(){
-  if (cst.PM2_LOG_DATE_FORMAT && typeof cst.PM2_LOG_DATE_FORMAT == 'string'){
-    var moment = require('moment');
-
-    // Generate timestamp prefix
-    function timestamp(){
-      return moment().format(cst.PM2_LOG_DATE_FORMAT) + ': ';
-    }
-
-    var hacks = ['info', 'log', 'error', 'warn'], consoled = {};
-
-    // store console functions.
-    hacks.forEach(function(method){
-      consoled[method] = console[method];
-    });
-
-    // Hack Console.
-    hacks.forEach(function(k){
-      console[k] = function(){
-        // do not destroy variable insertion
-        arguments[0] && (arguments[0] = timestamp() + arguments[0]);
-        consoled[k].apply(console, arguments);
-      };
-    });
-  }
-})();
-
-// process.on('uncaughtException', function(err) {
-//   if (err && err.message == 'Resource leak detected.') {
-//     // Catch and ignore this error
-//     // Throw by cluster module with Node 0.11.13<=
-//     console.error(err.stack);
-//     console.error('Resource leak detected for cluster module');
-//   }
-//   else if (err) {
-//     console.error(err.stack);
-
-//     if (God.dumpProcessList)
-//       God.dumpProcessList(function() {
-//         return process.exit(cst.ERROR_EXIT);
-//       });
-//     else
-//       return process.exit(cst.ERROR_EXIT);
-//   }
-// });
+Utility.overiddeConsole();
 
 /**
  * Populate God namespace
@@ -182,9 +113,10 @@ God.executeApp = function executeApp(env, cb) {
 
   //if (env_copy['command']) env_copy['command'].locked = false;
 
-  // Until Joyent accepts our PR, we have to do this trick
-  // to avoid `Resource leak error` due to 'disconnect' event
-  // not being fired sometimes
+  /**
+   * Avoid `Resource leak error` due to 'disconnect' event
+   * not being fired sometimes
+   */
   var workAround = function(worker) {
     var listeners = null;
