+++ lib/Utility.js
@@ -26,6 +26,32 @@ var Utility = module.exports = {
     } catch (e) {}
     return ret;
   },
+  overiddeConsole : function() {
+    if (cst.PM2_LOG_DATE_FORMAT && typeof cst.PM2_LOG_DATE_FORMAT == 'string'){
+      var moment = require('moment');
+
+      // Generate timestamp prefix
+      function timestamp(){
+        return moment().format(cst.PM2_LOG_DATE_FORMAT) + ': ';
+      }
+
+      var hacks = ['info', 'log', 'error', 'warn'], consoled = {};
+
+      // store console functions.
+      hacks.forEach(function(method){
+        consoled[method] = console[method];
+      });
+
+      // Hack Console.
+      hacks.forEach(function(k){
+        console[k] = function(){
+          // do not destroy variable insertion
+          arguments[0] && (arguments[0] = timestamp() + arguments[0]);
+          consoled[k].apply(console, arguments);
+        };
+      });
+    }
+  },
   startLogging : function(stds, callback) {
     /**
      * Start log outgoing messages
